separated logic for bundles type
