+++ bundle.go
@@ -1,9 +1,7 @@
 package versionbundle
 
 import (
-	"encoding/json"
 	"reflect"
-	"sort"
 	"strings"
 	"time"
 
@@ -109,78 +107,6 @@ func (b Bundle) Validate() error {
 	}
 
 	return nil
-}
-
-// ValidateBundles is a plain validation type for a list of version bundles. A
-// list of version bundles is exposed by authorities. Lists of version bundles
-// of multiple authorities are aggregated and grouped to reflect distributions.
-type ValidateBundles []Bundle
-
-func (b ValidateBundles) Copy() ValidateBundles {
-	raw, err := json.Marshal(b)
-	if err != nil {
-		panic(err)
-	}
-
-	var copy ValidateBundles
-	err = json.Unmarshal(raw, &copy)
-	if err != nil {
-		panic(err)
-	}
-
-	return copy
-}
-
-func (b ValidateBundles) Validate() error {
-	if len(b) == 0 {
-		return microerror.Maskf(invalidBundlesError, "version bundles must not be empty")
-	}
-
-	if b.hasDuplicatedVersions() {
-		return microerror.Maskf(invalidBundlesError, "version bundle versions must be unique")
-	}
-
-	b1 := b.Copy()
-	b2 := b.Copy()
-	sort.Sort(SortBundlesByTime(b1))
-	sort.Sort(SortBundlesByVersion(b2))
-	if !reflect.DeepEqual(b1, b2) {
-		return microerror.Maskf(invalidBundlesError, "version bundle versions must always increment")
-	}
-
-	for _, bundle := range b {
-		err := bundle.Validate()
-		if err != nil {
-			return microerror.Maskf(invalidBundlesError, err.Error())
-		}
-	}
-
-	bundleName := b[0].Name
-	for _, bundle := range b {
-		if bundle.Name != bundleName {
-			return microerror.Maskf(invalidBundlesError, "name must be the same for all version bundles")
-		}
-	}
-
-	return nil
-}
-
-func (b ValidateBundles) hasDuplicatedVersions() bool {
-	for _, b1 := range b {
-		var seen int
-
-		for _, b2 := range b {
-			if b1.Version == b2.Version {
-				seen++
-
-				if seen >= 2 {
-					return true
-				}
-			}
-		}
-	}
-
-	return false
 }
 
 // ValidateAggregatedBundles is a plain validation type for aggregated lists of
